refactor(voice): the capability assembler stays a package; fold the minter subclass
The capability assembler is host composition rather than Electron, so it stays
in `@sidecar/voice`. What goes is the abstraction inside: a one-implementer
interface file and two subclasses whose bodies were a `super()` call and a
refusal table. `ServiceRealtimeCredentialMinter` is concrete and takes an
optional `authorization`, whose absence is what makes the introduction mint
send no header and read a 401 as a plain failure; two factory functions keep
the call sites' option objects. Both refusal tables become one, whose 429 is a
spent allowance whatever reason it names.

The trace exporter is a developer CLI, so it leaves `@sidecar/devtrace` for
`tools/trace-export`, taking the package's reach for the hosted tool catalog
with it. `TRACE_ENTRY_KIND` moves to the trace vocabulary door the exporter can
read without resolving the writer's file handling.
