+++ tools/trace-export/package.json
@@ -1,21 +1,18 @@
 {
-  "name": "@sidecar/devtrace",
+  "name": "@luke/trace-export",
   "version": "0.5.0",
   "private": true,
   "type": "module",
-  "exports": {
-    ".": "./src/index.js",
-    "./vocabulary": "./src/vocabulary.js"
-  },
-  "types": "./src/index.ts",
+  "description": "Turns a recorded development trace into the document the local unbox-ai viewer opens",
   "scripts": {
+    "export": "tsx src/cli.ts",
     "test": "tsx --test 'src/**/*.test.ts'",
     "typecheck": "tsc --project tsconfig.json"
   },
   "dependencies": {
     "@sidecar/brain": "workspace:*",
+    "@sidecar/devtrace": "workspace:*",
     "@sidecar/realtime": "workspace:*",
-    "@sidecar/runtime": "workspace:*",
     "@sidecar/wire": "workspace:*"
   },
   "devDependencies": {
